refactor(ui): extract feed/card.ts — FeedItemCard component
